# Conflicts:
#	bin/de/hpi/javaide/breakout/elements/Ball.class
#	bin/de/hpi/javaide/breakout/elements/ui/Score.class
#	bin/de/hpi/javaide/breakout/screens/GameScreen.class
#	bin/de/hpi/javaide/breakout/screens/Screen.class
#	bin/de/hpi/javaide/breakout/screens/StartScreen.class
#	bin/de/hpi/javaide/breakout/starter/Game.class

+++ src/de/hpi/javaide/breakout/elements/Ball.java
@@ -52,7 +52,7 @@ public class Ball extends Elliptic {
 		direction.setY(-direction.getY());
 	}	
 	
-	public void setSpeed(int speed){
+	public void setSpeed(float speed){
 		direction.mult(speed);
 //		direction.normalize();
 	}
